+++ src/main/java/com/medipol/h5210017/midterm/service/number/NumberService.java
@@ -1,5 +1,5 @@
 package com.medipol.h5210017.midterm.service.number;
 
 public interface NumberService {
-    String reverse(int number);
+    String reverse(long number);
 }
